==> search_engine/data/3_term_frequency.ipynb <==
import pandas as pd
import math
import json

df = pd.read_csv('gcore_article_cut.csv', encoding='gb18030')
df = df.set_index('url')
df = df.fillna("")

### title + content

index = {}
for url, row in df.iterrows():
    index[url] = {}
    # 这里把标题内出现的词算作出现了5次，提高标题的权重
    for word in row['title'].split(' '):
        if word not in index[url]:
            index[url][word] = 5
        else:
            index[url][word] += 5
    for word in row['content'].split(' '):
        if word not in index[url]:
            index[url][word] = 1
        else:
            index[url][word] += 1
    if index[url].get('')!= None:
        del index[url]['']

list(index.items())[0]

tf = {}
for url, words in index.items():
    url_tf = {}
    for word, freq in words.items():
        url_tf[word] = 1 + math.log2(freq)
    tf[url] = url_tf

list(tf.items())[0]

# 所有文档内所有词的集合
word_set = []
for url, words in index.items():
    for word, freq in words.items():
        word_set.append(word)
word_set = sorted(set(word_set)) # 比起添加时筛选去重，最后统一去重的速度要快得多

inverted_index = {}
for url, words in index.items():
    for word, freq in words.items():
        if word not in inverted_index:
            inverted_index[word] = {}
        inverted_index[word][url] = freq

list(inverted_index.items())[0]

idf = {}
for url, words in index.items():
    for word, freq in words.items():
        if word not in idf:
            idf[word] = math.log2(len(index)/len(inverted_index[word]))

list(idf.items())[:5]

tf_idf = {}
for url, words in index.items():
    url_tf_idf = {}
    for word, freq in words.items():
        url_tf_idf[word] = tf[url][word] * idf[word]
    tf_idf[url] = url_tf_idf

list(tf_idf.items())[0]

# 提前离线计算每篇文档的向量长度，后面向量空间检索时可以直接用
doc_len = {}
for url, tf_idf_dict in tf_idf.items():
    tmp_len = 0
    for word, tf_idf_val in tf_idf_dict.items():
        tmp_len += pow(tf_idf_val,2)
    doc_len[url] = math.sqrt(tmp_len)

list(doc_len.items())[:5]

with open('../freq/word_set.json', 'w', encoding='utf-8') as f:
    json.dump(word_set, f, ensure_ascii=False)
with open('../freq/idf.json', 'w', encoding='utf-8') as f:
    json.dump(idf, f, ensure_ascii=False)
with open('../freq/tf-idf.json', 'w', encoding='utf-8') as f:
    json.dump(tf_idf, f, ensure_ascii=False)
with open('../freq/doc_len.json', 'w', encoding='utf-8') as f:
    json.dump(doc_len, f, ensure_ascii=False)

### title

index_title = {}
for url, row in df.iterrows():
    index_title[url] = {}
    for word in row['title'].split(' '):
        if word not in index_title[url]:
            index_title[url][word] = 5
        else:
            index_title[url][word] += 5
    if index_title[url].get('')!= None:
        del index_title[url]['']

tf_title = {}
for url, words in index_title.items():
    url_tf = {}
    for word, freq in words.items():
        url_tf[word] = 1 + math.log2(freq)
    tf_title[url] = url_tf

word_set_title = []
for url, words in index_title.items():
    for word, freq in words.items():
        word_set_title.append(word)
word_set_title = sorted(set(word_set_title))

inverted_index_title = {}
for url, words in index_title.items():
    for word, freq in words.items():
        if word not in inverted_index_title:
            inverted_index_title[word] = {}
        inverted_index_title[word][url] = freq

idf_title = {}
for url, words in index_title.items():
    for word, freq in words.items():
        if word not in idf_title:
            idf_title[word] = math.log2(len(index_title)/len(inverted_index_title[word]))

tf_idf_title = {}
for url, words in index_title.items():
    url_tf_idf = {}
    for word, freq in words.items():
        url_tf_idf[word] = tf_title[url][word] * idf_title[word]
    tf_idf_title[url] = url_tf_idf

doc_len_title = {}
for url, tf_idf_dict in tf_idf_title.items():
    tmp_len = 0
    for word, tf_idf_val in tf_idf_dict.items():
        tmp_len += pow(tf_idf_val,2)
    doc_len_title[url] = math.sqrt(tmp_len)

with open('../freq/word_set_title.json', 'w', encoding='utf-8') as f:
    json.dump(word_set_title, f, ensure_ascii=False)
with open('../freq/idf_title.json', 'w', encoding='utf-8') as f:
    json.dump(idf_title, f, ensure_ascii=False)
with open('../freq/tf-idf_title.json', 'w', encoding='utf-8') as f:
    json.dump(tf_idf_title, f, ensure_ascii=False)
with open('../freq/doc_len_title.json', 'w', encoding='utf-8') as f:
    json.dump(doc_len_title, f, ensure_ascii=False)
